Move common extension options into a separate file.
